+++ src/honeyhive/_generated/models/MetricVersionContentFilters.py
@@ -1,0 +1,22 @@
+from typing import Any, Dict, List, Optional, Union
+
+from pydantic import BaseModel, Field
+
+from .FiltersArray import FiltersArray
+
+__all__ = ["MetricVersionContentFilters"]
+
+
+class MetricVersionContentFilters(BaseModel):
+    """
+    MetricVersionContentFilters model
+    """
+
+    model_config = {
+        "populate_by_name": True,
+        "validate_assignment": True,
+        "extra": "allow",
+        "protected_namespaces": (),
+    }
+
+    filterArray: FiltersArray = Field(validation_alias="filterArray")